Safari-611.1.10/LayoutTests/fast/forms/select/menulist-oninput-fired.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<script src="../../../resources/js-test.js"></script>
</head>
<body>
<p>Selecting menu list options should fire input event</p>

<select id="bug-test" onInput="debug('PASS')">
  <option>1</option>
  <option>2</option>
  <option>3</option>
  <option>4</option>
  <option>5</option>
</select>
<div id="console"></div>
<script>
    document.getElementById("bug-test").focus();
    if (window.eventSender) {

        debug("Pressing 2 key");
        eventSender.keyDown('2');
    } else {
        debug("Test manually by pressing option key. Should get PASS message");
    }
</script>
</body>
</html>
